Final Changes
Final Changes to Sprint 3
